Data.Binary.Orphans

Description

Provides orphan Binary instances for types in various packages:

aeson

scientific (prior to scientific-0.3.4.0)

semigroups

tagged

text (through text-binary, or text >= 1.2.1)

time

unordered-containers

vector (through vector-binary-instances)

Also there is Binary Fixed instance.

Class re-export

class Binary t where

The Binary class provides put and get, methods to encode and decode a Haskell value to a lazy ByteString. It mirrors the Read and Show classes for textual representation of Haskell types, and is suitable for serialising Haskell values to disk, over the network.

For decoding and generating simple external binary formats (e.g. C structures), Binary may be used, but in general is not suitable for complex protocols. Instead use the Put and Get primitives directly.

Instances of Binary should satisfy the following property:

decode . encode == id

That is, the get and put methods should be the inverse of each other. A range of instances are provided for basic Haskell types.

Methods

put :: t -> Put

Encode a value in the Put monad.

get :: Get t

Decode a value in the Get monad

putList :: [t] -> Put

Encode a list of values in the Put monad. The default implementation may be overridden to be more efficient but must still have the same encoding format.

Module re-export

module Data.Binary

Orphan instances

Binary UTCTime

Methods

put :: UTCTime -> Put

get :: Get UTCTime

putList :: [UTCTime] -> Put

Binary Value

Methods

put :: Value -> Put

get :: Get Value

putList :: [Value] -> Put

Binary ZonedTime
Since: binary-orphans-0.1.8.0

Methods

put :: ZonedTime -> Put

get :: Get ZonedTime

putList :: [ZonedTime] -> Put

Binary LocalTime

Methods

put :: LocalTime -> Put

get :: Get LocalTime

putList :: [LocalTime] -> Put

Binary TimeOfDay

Methods

put :: TimeOfDay -> Put

get :: Get TimeOfDay

putList :: [TimeOfDay] -> Put

Binary TimeZone

Methods

put :: TimeZone -> Put

get :: Get TimeZone

putList :: [TimeZone] -> Put

Binary UniversalTime

Methods

put :: UniversalTime -> Put

get :: Get UniversalTime

putList :: [UniversalTime] -> Put

Binary SystemTime
Since: binary-orphans-0.1.7.0

Methods

put :: SystemTime -> Put

get :: Get SystemTime

putList :: [SystemTime] -> Put

Binary NominalDiffTime

Methods

put :: NominalDiffTime -> Put

get :: Get NominalDiffTime

putList :: [NominalDiffTime] -> Put

Binary AbsoluteTime
Since: binary-orphans-0.1.4.0

Methods

put :: AbsoluteTime -> Put

get :: Get AbsoluteTime

putList :: [AbsoluteTime] -> Put

Binary DiffTime

Methods

put :: DiffTime -> Put

get :: Get DiffTime

putList :: [DiffTime] -> Put

Binary Day

Methods

put :: Day -> Put

get :: Get Day

putList :: [Day] -> Put

(FoldCase a, Binary a) => Binary (CI a)
Since: binary-orphans-0.1.5.0

Methods

put :: CI a -> Put

get :: Get (CI a)

putList :: [CI a] -> Put

(Hashable a, Binary a) => Binary (Hashed a)

Methods

put :: Hashed a -> Put

get :: Get (Hashed a)

putList :: [Hashed a] -> Put

(Hashable v, Eq v, Binary v) => Binary (HashSet v)

Methods

put :: HashSet v -> Put

get :: Get (HashSet v)

putList :: [HashSet v] -> Put

(Hashable k, Eq k, Binary k, Binary v) => Binary (HashMap k v)

Methods

put :: HashMap k v -> Put

get :: Get (HashMap k v)

putList :: [HashMap k v] -> Put

Binary b => Binary (Tagged k s b)

Methods

put :: Tagged k s b -> Put

get :: Get (Tagged k s b)

putList :: [Tagged k s b] -> Put